Improve figure printing from the Debugger
When printing multiple viewports the contents of each must be clipped
to prevent them from spilling over.  Although GLPS supports this, it
seems to be buggy for PDF output.  We therefore switch back to using
EPS files for figures and let Latex convert them to PDFs, as required.

We also add support for printing figures as stippled outlines.